python script

+++ src/wavegen.py
@@ -24,7 +24,7 @@ def char_to_int(character):
     return unsigned_to_signed(ord(character))
 
 for i in range(100):
-    sig = file_in.readline()
+    sig = (file_in.readline()).astype(int)
     ser.write(int_to_char(int(sig)))
     d = ser.read()
     print(char_to_int(d))
